[feat] updated subscription methods
