+++ website/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@headlessui/react": "^2.1.0",
     "@supabase/supabase-js": "^2.46.2",
-    "@supabase/supabase-js": "^2.46.2",
     "@types/node": "^20.10.8",
     "@types/react": "^18.2.47",
     "@types/react-dom": "^18.2.18",
@@ -23,7 +22,6 @@
     "i18next-http-backend": "^3.0.1",
     "dotenv": "^16.4.7",
     "next": "^14.0.4",
-    "postmark": "^4.0.5",
     "postmark": "^4.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
