+++ website/src/components/Features.tsx
@@ -2,6 +2,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { CheckIcon } from './CheckIcon';
+import i18next from '@/i18n';
 import Image from 'next/image';
 import sampleImage from '@/images/sampleImage.png';
 import { Button } from '@/components/Button';
@@ -12,23 +13,19 @@ export function Features() {
   const { t } = useTranslation();
   const [email, setEmail] = useState('');
   const [message, setMessage] = useState('');
-  const [isSubmitting, setIsSubmitting] = useState(false);
 
   const handleEmailSubmission = async (e: React.FormEvent) => {
     e.preventDefault();
-    setIsSubmitting(true);
     setMessage('');
 
     try {
       await handleSubscription(email);
 
-      await handleSendWelcomeEmail(email);
+      //await handleSendWelcomeEmail(email);
 
       setMessage(t('feature.successEmail', { email }));
     } catch (error: any) {
       setMessage(error.message || t('feature.genericError'));
-    } finally {
-      setIsSubmitting(false);
     }
   };
 
@@ -62,7 +59,7 @@ export function Features() {
               <h3 className="text-lg font-medium tracking-tight text-black">
                 {t('feature.try')}
               </h3>
-              <div className="mt-4 sm:relative sm:flex sm:items-center sm:py-0.5 sm:pr-2.5">
+              <div className="mt-4 items-center flex inset-0 rounded-md border border-black/40 peer-focus:border-black peer-focus:ring-1 peer-focus:ring-blue-300 sm:rounded-xl">
                 <div className="relative sm:static sm:flex-auto">
                   <input
                     type="email"
@@ -74,13 +71,13 @@ export function Features() {
                     onChange={(e) => setEmail(e.target.value)} // Mettez à jour l'email dans l'état
                     className="peer relative z-10 w-full appearance-none bg-transparent px-4 py-2 text-base text-black placeholder:text-slate-400/70 focus:outline-none sm:py-3"
                   />
-                  <div className="absolute inset-0 rounded-md border border-black/40 peer-focus:border-black peer-focus:ring-1 peer-focus:ring-blue-300 sm:rounded-xl" />
+                  {/* <div className="absolute " /> */}
                 </div>
                 <Button
                   type="submit"
                   variant="solid"
                   color="blue"
-                  className="mt-4 w-full sm:relative sm:z-10 sm:mt-0 sm:w-auto sm:flex-none"
+                  className=" w-full m-2 sm:relative sm:z-10 sm:w-auto sm:flex-none"
                 >
                   {t('feature.sub')}
                 </Button>
