+++ website/src/locales/fr.json
@@ -24,5 +24,9 @@
   "footer": {
     "copyright": "Copyright &copy; {{date}} Marmelab",
     "allRightsReserved": "Tous droits réservés."
+  },
+  "mail": {
+    "txt": "Bienvenue chez Curator AI ! Curator AI est là pour simplifier votre suivi de l'information. Avec notre IA, vous recevrez une newsletter contenant des articles résumés et sélectionnés en fonction de vos sujets et sources préférés. Envoyez un mail avec l'une de ces instructions pour intéragir avec Curator AI :\n### 📚 Ajouter des sujets : *Je veux suivre les sujets IA et santé.*\n### 🌐 Ajouter des sites : *Ajouter https://www.lemonde.fr/ à mes sources.*\n### ⏰ Définir la périodicité : *Envoyez-moi la newsletter tous les lundis à 8h.*\nPar défaut, vous recevrez votre première newsletter dans les 5 minutes, puis une fois par semaine le même jour et à la même heure que votre inscription.",
+    "html": "<!DOCTYPE html><html><head>STYLE TOKEN</head><body><div class=\"email-container\"><h1 class=\"header\">Bienvenue chez Curator AI!</h1><p class=\"intro\">Nous sommes ravis de vous accueillir 🎉.</p><p class=\"content\">Avec notre IA, vous recevrez une newsletter personnalisée contenant les articles les plus pertinents, résumés et adaptés à vos préférences.</p><h2 class=\"subheader\">Comment interagir avec Curator AI ? Envoyez un mail avec l'une de ces instructions :</h2><ul class=\"instructions\"><li> 📚 Ajouter des sujets : <strong>\"Je veux suivre les sujets IA et santé.\"</strong></li><li> 🌐 Ajouter un site : <strong>\"Ajouter https://www.lemonde.fr/ à mes sources.\"</strong></li><li> ⏰ Définir la périodicité : <strong>\"Envoyez-moi la newsletter tous les lundis à 8h.\"</strong></li></ul><p class=\"reminder\">📩 Votre première newsletter sera envoyée dans les 5 minutes suivant votre inscription. Par défaut, les futures newsletters arriveront chaque semaine à la même heure que votre inscription.</p><p class=\"closing\">Pour toute question ou modification, répondez simplement à cet email. Bonne exploration!</p><p class=\"signature\">L'équipe Curator AI 🚀</p></div></body></html>"
   }
 }
