+++ website/src/services/subscribeService.ts
@@ -1,4 +1,9 @@
-// src/services/subscribeService.ts
+import { createClient } from '@supabase/supabase-js';
+
+// Initialize Supabase with environment variables
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!;
+const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!;
+const supabase = createClient(supabaseUrl, supabaseAnonKey);
 
 import { validateEmail } from '@/utils/validateEmail'; // Import the validation utility
 
@@ -7,26 +12,39 @@ import { validateEmail } from '@/utils/validateEmail'; // Import the validation 
  * @param email - The email address to subscribe.
  */
 export async function handleSubscription(email: string): Promise<void> {
-  if (!validateEmail(email)) {
-    throw new Error('Invalid email address.');
-  }
 
-  try {
-    const response = await fetch('/api/subscribe', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ email }),
-    });
-
-    const result = await response.json();
-
-    if (!response.ok) {
-      throw new Error(result.error || 'Failed to subscribe the email.');
+    if (!validateEmail(email)) {
+      throw new Error('Invalid email address.');
     }
-  } catch (error) {
-    console.error('Error while subscribing email:', error);
-    throw error; // Re-throw the error for the caller to handle.
-  }
+  
+    try {
+      // Check if the email already exists in the "subscribers" table
+      const { data: existingEmail, error: selectError } = await supabase
+        .from('subscribers')
+        .select('user_email')
+        .eq('user_email', email)
+        .single();
+  
+      if (selectError && selectError.code !== 'PGRST116') {
+        throw new Error(`Error verifying email: ${selectError.message}`);
+      }
+  
+      if (existingEmail) { 
+        throw new Error('Email already registered.');
+      }
+  
+      // Insert the email into the "subscribers" table
+      const { error: insertError } = await supabase
+        .from('subscribers')
+        .insert([{ user_email: email }]);
+  
+      if (insertError) {
+        throw new Error(`Error inserting email: ${insertError.message}`);
+      }
+  
+      console.log('Email successfully registered.');
+    } catch (error) {
+      console.error('Unexpected error:', error);
+      throw error;
+    }
 }
